requirements.txt
# Core AI and ML libraries
openai>=1.0.0
autogen-agentchat>=0.2.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# Environment variables
python-dotenv>=1.0.0

# Data processing and validation
pydantic>=2.0.0
jsonschema>=4.19.0
python-dateutil>=2.8.0

# Logging
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development tools
black>=23.7.0
flake8>=6.0.0